Fix pinmap type, must be byte sized

+++ Arduino/pin_map.h
@@ -5,6 +5,6 @@
 
 #define PIN_BUTTON (4)
 #define PIN_D_BUS (5)
-const uint PIN_LEDS[] = {16, 0, 2, 14, 12, 13};
+const uint8_t PIN_LEDS[] = {16, 0, 2, 14, 12, 13};
 
 #endif
